+++ .editorconfig
@@ -1,0 +1,8 @@
+[*.xml]
+indent_style = space
+indent_size = 2
+
+[*.java]
+indent_style = space
+indent_size = 4
+
