+++ .gitignore
@@ -1,0 +1,3 @@
+# Ignore mvn artifacts
+/target
+
