--- release/pdfsign
@@ -1,2 +1,0 @@
-cd /home/paxy/sh/
-java -jar pdfsign.jar "$1"
