--- release/pkcs11-posta.cfg
@@ -1,2 +1,0 @@
-name = Posta
-library = /usr/lib/libaetpkss.so.3
